Import Basic Version
This is the basic version of the methods - the "functions" folder is all the functions required for count variables (via poisson) and "function binomial" is for rate variables (via binomial)

As things stand, there is a lot left to do, but this is a very minimal version which *does* work.

Currently the ridge regularisation doesn't work. The code for it is there, but it estimates lambda used in ridge regression via the likelihood rather than the leave-one-out cross validated likelihood) so the functioning version just ignores the ridge regression completely.